basic auth

+++ src/main/java/com/hotelbooking/dto/ReviewDto.java
@@ -1,12 +1,17 @@
 package com.hotelbooking.dto;
 
+import jakarta.validation.constraints.DecimalMax;
+import jakarta.validation.constraints.DecimalMin;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class ReviewDto {
     private Long id;
-    private Integer rating;
+
+    private Double rating;
     private String comment;
     private LocalDate reviewDate;
     private Long userId;
